pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "djangocms-events"
version = "0.1.3"
authors = [
  { name="Jens-Erik Weber", email="web@passiv.de" },
]
description = "Events list and calendar for django CMS"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE.txt"}
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "Django>=3.2",
    "django-cms>=3.11",
    "easy-thumbnails>=2.8",
    "django-quill-editor>=0.1",
    "ics>=0.7",
    "requests>=2.28",
    "arrow>=1.2",
    "django-filer>=2.2",
]

[project.urls]
Homepage = "https://github.com/MacLake/djangocms-events"
repository = "https://github.com/MacLake/djangocms-events.git"

[tool.yapf]
based_on_style = "facebook"
SPLIT_BEFORE_BITWISE_OPERATOR = true
SPLIT_BEFORE_ARITHMETIC_OPERATOR = true
SPLIT_BEFORE_LOGICAL_OPERATOR = true
SPLIT_BEFORE_DOT = true
